Pull out a dedicated status report file
