+++ lib/actions-util.test.js
@@ -213,37 +213,4 @@ const util_1 = require("./util");
         getAdditionalInputStub.restore();
     });
 });
-(0, ava_1.default)("createStatusReportBase", async (t) => {
-    await (0, util_1.withTmpDir)(async (tmpDir) => {
-        (0, testing_utils_1.setupActionsVars)(tmpDir, tmpDir);
-        process.env["GITHUB_REF"] = "refs/heads/main";
-        process.env["GITHUB_SHA"] = "a".repeat(40);
-        process.env["GITHUB_RUN_ID"] = "100";
-        process.env["GITHUB_RUN_ATTEMPT"] = "2";
-        process.env["GITHUB_REPOSITORY"] = "octocat/HelloWorld";
-        process.env["CODEQL_ACTION_ANALYSIS_KEY"] = "analysis-key";
-        process.env["RUNNER_OS"] = "macOS";
-        const getRequiredInput = sinon.stub(actionsUtil, "getRequiredInput");
-        getRequiredInput.withArgs("matrix").resolves("input/matrix");
-        const statusReport = await (0, api_client_1.createStatusReportBase)("init", "failure", new Date("May 19, 2023 05:19:00"), "failure cause", "exception stack trace");
-        t.assert(typeof statusReport.job_run_uuid === "string");
-        t.assert(statusReport.workflow_run_id === 100);
-        t.assert(statusReport.workflow_run_attempt === 2);
-        t.assert(statusReport.workflow_name === (process.env["GITHUB_WORKFLOW"] || ""));
-        t.assert(statusReport.job_name === (process.env["GITHUB_JOB"] || ""));
-        t.assert(statusReport.analysis_key === "analysis-key");
-        t.assert(statusReport.commit_oid === process.env["GITHUB_SHA"]);
-        t.assert(statusReport.ref === process.env["GITHUB_REF"]);
-        t.assert(statusReport.action_name === "init");
-        t.assert(statusReport.action_oid === "unknown");
-        t.assert(statusReport.started_at === process.env[environment_1.EnvVar.WORKFLOW_STARTED_AT]);
-        t.assert(statusReport.action_started_at ===
-            new Date("May 19, 2023 05:19:00").toISOString());
-        t.assert(statusReport.status === "failure");
-        t.assert(statusReport.cause === "failure cause");
-        t.assert(statusReport.exception === "exception stack trace");
-        t.assert(statusReport.runner_os === process.env["RUNNER_OS"]);
-        t.assert(typeof statusReport.action_version === "string");
-    });
-});
 //# sourceMappingURL=actions-util.test.js.map
