+++ lib/analyze-action-input.test.js
@@ -32,6 +32,7 @@ const actionsUtil = __importStar(require("./actions-util"));
 const analyze = __importStar(require("./analyze"));
 const api = __importStar(require("./api-client"));
 const configUtils = __importStar(require("./config-utils"));
+const statusReport = __importStar(require("./status-report"));
 const testing_utils_1 = require("./testing-utils");
 const util = __importStar(require("./util"));
 (0, testing_utils_1.setupTests)(ava_1.default);
@@ -47,9 +48,9 @@ const util = __importStar(require("./util"));
         process.env["GITHUB_REPOSITORY"] = "github/codeql-action-fake-repository";
         process.env["GITHUB_API_URL"] = "https://api.github.com";
         sinon
-            .stub(api, "createStatusReportBase")
+            .stub(statusReport, "createStatusReportBase")
             .resolves({});
-        sinon.stub(api, "sendStatusReport").resolves(true);
+        sinon.stub(statusReport, "sendStatusReport").resolves(true);
         const gitHubVersion = {
             type: util.GitHubVariant.DOTCOM,
         };
