+++ lib/analyze-action.js
@@ -34,7 +34,6 @@ const core = __importStar(require("@actions/core"));
 const actionsUtil = __importStar(require("./actions-util"));
 const analyze_1 = require("./analyze");
 const api_client_1 = require("./api-client");
-const api = __importStar(require("./api-client"));
 const autobuild_1 = require("./autobuild");
 const codeql_1 = require("./codeql");
 const config_utils_1 = require("./config-utils");
@@ -44,14 +43,16 @@ const feature_flags_1 = require("./feature-flags");
 const languages_1 = require("./languages");
 const logging_1 = require("./logging");
 const repository_1 = require("./repository");
+const statusReport = __importStar(require("./status-report"));
+const status_report_1 = require("./status-report");
 const trap_caching_1 = require("./trap-caching");
 const uploadLib = __importStar(require("./upload-lib"));
 const util = __importStar(require("./util"));
 const util_1 = require("./util");
 async function sendStatusReport(startedAt, config, stats, error, trapCacheUploadTime, dbCreationTimings, didUploadTrapCaches, logger) {
-    const status = actionsUtil.getActionsStatus(error, stats?.analyze_failure_language);
-    const statusReportBase = await api.createStatusReportBase("finish", status, startedAt, error?.message, error?.stack);
-    const statusReport = {
+    const status = (0, status_report_1.getActionsStatus)(error, stats?.analyze_failure_language);
+    const statusReportBase = await (0, status_report_1.createStatusReportBase)("finish", status, startedAt, error?.message, error?.stack);
+    const report = {
         ...statusReportBase,
         ...(config
             ? {
@@ -63,14 +64,14 @@ async function sendStatusReport(startedAt, config, stats, error, trapCacheUpload
     };
     if (config && didUploadTrapCaches) {
         const trapCacheUploadStatusReport = {
-            ...statusReport,
+            ...report,
             trap_cache_upload_duration_ms: Math.round(trapCacheUploadTime || 0),
             trap_cache_upload_size_bytes: Math.round(await (0, trap_caching_1.getTotalCacheSize)(config.trapCaches, logger)),
         };
-        await api.sendStatusReport(trapCacheUploadStatusReport);
+        await statusReport.sendStatusReport(trapCacheUploadStatusReport);
     }
     else {
-        await api.sendStatusReport(statusReport);
+        await statusReport.sendStatusReport(report);
     }
 }
 exports.sendStatusReport = sendStatusReport;
@@ -146,7 +147,7 @@ async function run() {
     util.initializeEnvironment(actionsUtil.getActionVersion());
     const logger = (0, logging_1.getActionsLogger)();
     try {
-        if (!(await api.sendStatusReport(await api.createStatusReportBase("finish", "starting", startedAt)))) {
+        if (!(await statusReport.sendStatusReport(await (0, status_report_1.createStatusReportBase)("finish", "starting", startedAt)))) {
             return;
         }
         config = await (0, config_utils_1.getConfig)(actionsUtil.getTemporaryDirectory(), logger);
