+++ lib/init-action-post.js
@@ -36,6 +36,7 @@ const feature_flags_1 = require("./feature-flags");
 const initActionPostHelper = __importStar(require("./init-action-post-helper"));
 const logging_1 = require("./logging");
 const repository_1 = require("./repository");
+const status_report_1 = require("./status-report");
 const util_1 = require("./util");
 async function runWrapper() {
     const startedAt = new Date();
@@ -51,15 +52,15 @@ async function runWrapper() {
     catch (unwrappedError) {
         const error = (0, util_1.wrapError)(unwrappedError);
         core.setFailed(error.message);
-        await (0, api_client_1.sendStatusReport)(await (0, api_client_1.createStatusReportBase)("init-post", (0, actions_util_1.getActionsStatus)(error), startedAt, error.message, error.stack));
+        await (0, status_report_1.sendStatusReport)(await (0, status_report_1.createStatusReportBase)("init-post", (0, status_report_1.getActionsStatus)(error), startedAt, error.message, error.stack));
         return;
     }
-    const statusReportBase = await (0, api_client_1.createStatusReportBase)("init-post", "success", startedAt);
+    const statusReportBase = await (0, status_report_1.createStatusReportBase)("init-post", "success", startedAt);
     const statusReport = {
         ...statusReportBase,
         ...uploadFailedSarifResult,
     };
-    await (0, api_client_1.sendStatusReport)(statusReport);
+    await (0, status_report_1.sendStatusReport)(statusReport);
 }
 void runWrapper();
 //# sourceMappingURL=init-action-post.js.map
