+++ lib/upload-sarif-action.js
@@ -26,23 +26,23 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const core = __importStar(require("@actions/core"));
 const actionsUtil = __importStar(require("./actions-util"));
 const actions_util_1 = require("./actions-util");
-const api_client_1 = require("./api-client");
 const logging_1 = require("./logging");
 const repository_1 = require("./repository");
+const status_report_1 = require("./status-report");
 const upload_lib = __importStar(require("./upload-lib"));
 const util_1 = require("./util");
 async function sendSuccessStatusReport(startedAt, uploadStats) {
-    const statusReportBase = await (0, api_client_1.createStatusReportBase)("upload-sarif", "success", startedAt);
+    const statusReportBase = await (0, status_report_1.createStatusReportBase)("upload-sarif", "success", startedAt);
     const statusReport = {
         ...statusReportBase,
         ...uploadStats,
     };
-    await (0, api_client_1.sendStatusReport)(statusReport);
+    await (0, status_report_1.sendStatusReport)(statusReport);
 }
 async function run() {
     const startedAt = new Date();
     (0, util_1.initializeEnvironment)((0, actions_util_1.getActionVersion)());
-    if (!(await (0, api_client_1.sendStatusReport)(await (0, api_client_1.createStatusReportBase)("upload-sarif", "starting", startedAt)))) {
+    if (!(await (0, status_report_1.sendStatusReport)(await (0, status_report_1.createStatusReportBase)("upload-sarif", "starting", startedAt)))) {
         return;
     }
     try {
@@ -62,7 +62,7 @@ async function run() {
         const message = error.message;
         core.setFailed(message);
         console.log(error);
-        await (0, api_client_1.sendStatusReport)(await (0, api_client_1.createStatusReportBase)("upload-sarif", actionsUtil.getActionsStatus(error), startedAt, message, error.stack));
+        await (0, status_report_1.sendStatusReport)(await (0, status_report_1.createStatusReportBase)("upload-sarif", (0, status_report_1.getActionsStatus)(error), startedAt, message, error.stack));
         return;
     }
 }
