+++ src/actions-util.ts
@@ -239,126 +239,6 @@ function getRefFromEnv(): string {
   return refEnv;
 }
 
-export type ActionName =
-  | "init"
-  | "autobuild"
-  | "finish"
-  | "upload-sarif"
-  | "init-post"
-  | "resolve-environment";
-export type ActionStatus =
-  | "starting"
-  | "aborted"
-  | "success"
-  | "failure"
-  | "user-error";
-
-// Any status report may include an array of EventReports associated with it.
-export interface EventReport {
-  /** An enumerable description of the event. */
-  event: string;
-  /** Time this event started. */
-  started_at: string;
-  /** Time this event ended. */
-  completed_at: string;
-  /** eg: `success`, `failure`, `timeout`, etc. */
-  exit_status?: string;
-  /** If the event is language-specific. */
-  language?: string;
-  /**
-   * A generic JSON blob of data related to this event.
-   * Use Object.assign() to append additional fields to the object.
-   */
-  properties?: object;
-}
-
-export interface StatusReportBase {
-  /**
-   * UUID representing the job run that this status report belongs to. We
-   * generate our own UUID here because Actions currently does not expose a
-   * unique job run identifier. This UUID will allow us to more easily match
-   * reports from different steps in the same workflow job.
-   *
-   * If and when Actions does expose a unique job ID, we plan to populate a
-   * separate int field, `job_run_id`, with the Actions-generated identifier,
-   * as it will allow us to more easily join our telemetry data with Actions
-   * telemetry tables.
-   */
-  job_run_uuid: string;
-  /** ID of the workflow run containing the action run. */
-  workflow_run_id: number;
-  /** Attempt number of the run containing the action run. */
-  workflow_run_attempt: number;
-  /** Workflow name. Converted to analysis_name further down the pipeline.. */
-  workflow_name: string;
-  /** Job name from the workflow. */
-  job_name: string;
-  /** Analysis key, normally composed from the workflow path and job name. */
-  analysis_key: string;
-  /** Value of the matrix for this instantiation of the job. */
-  matrix_vars?: string;
-  /** Commit oid that the workflow was triggered on. */
-  commit_oid: string;
-  /** Ref that the workflow was triggered on. */
-  ref: string;
-  /** Name of the action being executed. */
-  action_name: ActionName;
-  /** Version of the action being executed, as a ref. */
-  action_ref?: string;
-  /** Version of the action being executed, as a commit oid. */
-  action_oid: string;
-  /** Time the first action started. Normally the init action. */
-  started_at: string;
-  /** Time this action started. */
-  action_started_at: string;
-  /** Time this action completed, or undefined if not yet completed. */
-  completed_at?: string;
-  /** State this action is currently in. */
-  status: ActionStatus;
-  /**
-   * Testing environment: Set if non-production environment.
-   * The server accepts one of the following values:
-   *  `["", "qa-rc", "qa-rc-1", "qa-rc-2", "qa-experiment-1", "qa-experiment-2", "qa-experiment-3"]`.
-   */
-  testing_environment: string;
-  /**
-   * Information about the enablement of the ML-powered JS query pack.
-   *
-   * @see {@link util.getMlPoweredJsQueriesStatus}
-   */
-  ml_powered_javascript_queries?: string;
-  /** Cause of the failure (or undefined if status is not failure). */
-  cause?: string;
-  /** Stack trace of the failure (or undefined if status is not failure). */
-  exception?: string;
-  /** Action runner operating system (context runner.os). */
-  runner_os: string;
-  /** Action runner hardware architecture (context runner.arch). */
-  runner_arch?: string;
-  /** Action runner operating system release (x.y.z from os.release()). */
-  runner_os_release?: string;
-  /** Action version (x.y.z from package.json). */
-  action_version: string;
-  /** CodeQL CLI version (x.y.z from the CLI). */
-  codeql_version?: string;
-}
-
-export interface DatabaseCreationTimings {
-  scanned_language_extraction_duration_ms?: number;
-  trap_import_duration_ms?: number;
-}
-
-export function getActionsStatus(
-  error?: unknown,
-  otherFailureCause?: string,
-): ActionStatus {
-  if (error || otherFailureCause) {
-    return error instanceof UserError ? "user-error" : "failure";
-  } else {
-    return "success";
-  }
-}
-
 export function getActionVersion(): string {
   return pkg.version!;
 }
