+++ src/analyze-action-env.test.ts
@@ -5,6 +5,7 @@ import * as actionsUtil from "./actions-util";
 import * as analyze from "./analyze";
 import * as api from "./api-client";
 import * as configUtils from "./config-utils";
+import * as statusReport from "./status-report";
 import {
   setupTests,
   setupActionsVars,
@@ -27,9 +28,9 @@ test("analyze action with RAM & threads from environment variables", async (t) =
     process.env["GITHUB_REPOSITORY"] = "github/codeql-action-fake-repository";
     process.env["GITHUB_API_URL"] = "https://api.github.com";
     sinon
-      .stub(api, "createStatusReportBase")
-      .resolves({} as actionsUtil.StatusReportBase);
-    sinon.stub(api, "sendStatusReport").resolves(true);
+      .stub(statusReport, "createStatusReportBase")
+      .resolves({} as statusReport.StatusReportBase);
+    sinon.stub(statusReport, "sendStatusReport").resolves(true);
     sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
 
     const gitHubVersion: util.GitHubVersion = {
