+++ src/init-action-post.ts
@@ -6,22 +6,19 @@
 
 import * as core from "@actions/core";
 
-import {
-  getActionsStatus,
-  getTemporaryDirectory,
-  printDebugLogs,
-  StatusReportBase,
-} from "./actions-util";
-import {
-  createStatusReportBase,
-  getGitHubVersion,
-  sendStatusReport,
-} from "./api-client";
+import { getTemporaryDirectory, printDebugLogs } from "./actions-util";
+import { getGitHubVersion } from "./api-client";
 import * as debugArtifacts from "./debug-artifacts";
 import { Features } from "./feature-flags";
 import * as initActionPostHelper from "./init-action-post-helper";
 import { getActionsLogger } from "./logging";
 import { parseRepositoryNwo } from "./repository";
+import {
+  StatusReportBase,
+  sendStatusReport,
+  createStatusReportBase,
+  getActionsStatus,
+} from "./status-report";
 import {
   checkGitHubVersionInRange,
   getRequiredEnvParam,
